+++ extras/integration_tests/web_wasm/lib/initialization_mode.dart
@@ -1,5 +1,6 @@
 enum InitializationMode {
   none,
+  noneAndDisableMigrations,
   loadAsset,
   migrateCustomWasmDatabase,
 }